pyproject.toml
[tool.poetry]
name = "rag-medical"
version = "0.1.0"
description = ""
authors = ["nguyen <nguyenhainguyen2606@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
llama-index = "^0.12.30"
fastapi = "^0.115.12"
faiss-cpu = "^1.10.0"
openai = "^1.73.0"
pytest = "^8.3.5"
ruff = "^0.11.5"
sqlmodel = "^0.0.24"
streamlit = "^1.44.1"
pyarrow = "^19.0.1"
uvicorn = "^0.34.0"
python-dotenv = "^1.1.0"
llama-index-vector-stores-faiss = "^0.3.0"
transformers = "^4.51.3"
scikit-learn = "^1.6.1"
tokenizers = "^0.21.1"
pydantic = "^2.11.3"
datasets = "^3.5.0"
tqdm = "^4.67.1"
matplotlib = "^3.10.1"
setuptools = "^78.1.0"
pytest-asyncio = "^0.26.0"
psycopg2 = "^2.9.10"
asyncpg = "^0.30.0"
watchdog = "^6.0.0"
numpy = "^2.2.5"
tomlkit = "^0.13.2"
passlib = "^1.7.4"
bcrypt = "<4.0"
loguru = "^0.7.3"
hf-xet = "^1.1.5"
whisperx = "^3.4.2"
silero-tts = "^0.0.5"
torch = "^2.7.1"
torchaudio = "^2.7.1"
silero = "^0.4.1"
streamlit-audiorec = "^0.1.3"
redis = "^6.2.0"
accelerate = "^1.9.0"
pandas = "^2.3.1"
openpyxl = "^3.1.5"
fpdf = "^1.7.2"



[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.1.1"
pytest-mock = "^3.14.0"
evaluate = "^0.4.3"
rouge-score = "^0.1.2"
bert-score = "^0.3.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
